tests: Add test for direct verb efadv_query_ah
Add test_query_efa_ah which use EFA's direct verb efadv_query_ah.
